Incorporate articles included in the review
These were screened on Rayyan and deemed worthy of inclusion in the
review so we will use the list to create a dataset that matches.

Note: some of the items that were screened have since been removed
from CGIAR (mostly CIFOR) repositories. For the handful I manually
checked I did not see CGIAR affiliations on the publisher pages. I
will leave them in the `data/included-in-review.csv` file for now.