src/merge_source_csvs.py: re-work combined dataset
We have one list of recorsd for the combined dataset now. In theory
it was better to have the primay list of DOIs from the Rayyan scre-
ening and the secondary list of reviews, syntheses, opinions, etc,
but it became trick to reconcile those. The list of DOIs for the
"combined" dataset corresponds to the DOIs that we considered final
on 2024-10-18.